machine: uart16550: Support Xilinx AXI UART16550
Add "xlnx,xps-uart16550-2.00.a" to the compatible list.

Check for these properties in the device tree node:
"reg-offset", "reg-shift", "clock-frequency", "current-speed"

Derive the clock divisor from the device tree, and do not hardcode the
line speed to 38400 bauds.
If the clock divisor cannot be derived from the device tree, leave it at
its boot setting.